feature-2 - add a project to generate allure rep. files
