+++ README.md
@@ -1,1 +1,8 @@
 # allure-reporter-test-jee
+
+## Generating reporting files with maven
+We can use a terminal:
+```
+mvn test
+``` 
+After the build is completed, the reporting files will be generated in the `allure-results` folder
